src/text_translation/gemini_flash/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Google AI SDK
google-generativeai==0.3.2

# Utilities
httpx==0.25.2
aiofiles==23.2.1
python-dotenv==1.0.0

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0

# Language detection (optional)
langdetect==1.0.9

# Performance monitoring
psutil==5.9.6